+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Network #19, upping the number of input buckets from 4 to 8 (2.95 +- 2.37)
 * The reverse futility pruning margin is adjusted if we are improving (2.68 +- 2.34)
 * Legal killer moves are now yielded before generating quiet moves for a minor speedup
+* A single stop flag is now used for all threads (5.69 +- 4.73 reg)
 
 #### Misc